Move solution picking options code to its own file.
